+++ pbp/json_generator/metadata_extractor.py
@@ -1,8 +1,7 @@
-# pypam-based-processing, Apache License 2.0
-# Filename: json_generator/wavfile.py
-# Description:  wav file metadata reader. Supports SoundTrap and icListen wav files
+# pbp, Apache License 2.0
+# Filename: json_generator/metadata_extractor.py
+# Description: Utilities for wav file metadata reading. Supports SoundTrap, NRS and and icListen audio files
 
-from logging import exception, warning, debug
 from pathlib import Path
 from typing import Optional
 
@@ -34,7 +33,7 @@ class AudioFile:
         self.frames = -1
         self.channels = -1
         self.subtype = ""
-        self.exception = np.NAN
+        self.exception = np.nan
 
     def has_exception(self):
         return True if len(self.exception) > 0 else False
@@ -121,15 +120,16 @@ class SoundTrapWavFile(AudioFile):
         self.frames = sample_count
         self.channels = 1
         self.subtype = "SoundTrap"
-        self.exception = np.NAN  # no exceptions for SoundTrap  files
+        self.exception = np.nan  # no exceptions for SoundTrap  files
 
 
 class IcListenWavFile(AudioFile):
     """IcListenWavFile uses the metadata from the wav file itself,
     but only grabs the needed metadata from the header in S3"""
 
-    def __init__(self, path_or_url: str, start: datetime):
+    def __init__(self, log, path_or_url: str, start: datetime):
         super().__init__(path_or_url, start)
+        self.log = log
         self.path_or_url = path_or_url
         self.start = start
         self.duration_secs = -1
@@ -137,7 +137,7 @@ class IcListenWavFile(AudioFile):
         self.frames = -1
         self.channels = -1
         self.subtype = ""
-        self.exception = np.NAN
+        self.exception = np.nan
         self.path_or_url = path_or_url
         bytes_per_sec = (
             3 * 256e3
@@ -169,20 +169,18 @@ class IcListenWavFile(AudioFile):
                     # header which is 44 bytes. Round the duration to the nearest second since the recording is
                     # always in 1 second increments
                     self.duration_secs = round(content_length - 44) / bytes_per_sec
-                    warning(self.exception)
+                    self.log.warning(self.exception)
             else:
                 info = sf.info(path_or_url)
                 self.duration_secs = info.duration
 
-            self.end = self.start + timedelta(
-                microseconds=int(info.frames * 1e6 / info.samplerate)
-            )
+            self.end = self.start + timedelta(microseconds=int(self.duration_secs * 1e6))
             self.fs = info.samplerate
             self.frames = info.frames
             self.channels = info.channels
             self.subtype = info.subtype if info.subtype else ""
         except Exception as ex:
-            exception(f"Corrupt file {path_or_url}. {ex}")
+            self.log.exception(f"Corrupt file {path_or_url}. {ex}")
 
 
 class FlacFile(AudioFile):
@@ -199,7 +197,7 @@ class FlacFile(AudioFile):
         self.frames = -1
         self.channels = -1
         self.subtype = ""
-        self.exception = np.NAN
+        self.exception = np.nan
         self.path_or_url = path_or_url
 
         try:
@@ -212,7 +210,6 @@ class FlacFile(AudioFile):
 
                 # get the duration from the extra_info data field which stores the duration in total bytes
                 fields = info.extra_info.split(":")
-                debug("\n".join(fields))
                 sample_rate = int(fields[3].split("\n")[0])
                 channels = int(fields[2].split("\n")[0])
                 length_microseconds = int(info.frames * 1e6 / info.samplerate)
@@ -222,7 +219,8 @@ class FlacFile(AudioFile):
                 # files are in the format NRS11_20191231_230836.flac'
                 # extract the timestamp from the file name
                 f = Path(file_name).stem.split("_")
-                # If the last two digits of the timestamp are 60, subtract 1 seconds
+                # If the last two digits of the timestamp are 60, subtract 1 second
+                # This is a bug in the FlacFile name
                 if f[2][-2:] == "60":
                     f = f[1] + f[2]
                     # Make the last two digits 59
@@ -248,4 +246,4 @@ class FlacFile(AudioFile):
                 self.channels = info.channels
                 self.subtype = info.subtype if info.subtype else ""
         except Exception as ex:
-            exception(f"Corrupt file {path_or_url}. {ex}")
+            self.log.exception(f"Corrupt file {path_or_url}. {ex}")