+++ pbp/json_generator/utils.py
@@ -1,5 +1,14 @@
+# pbp, Apache License 2.0
+# Filename: json_generator/utils.py
+# Description:  Utility functions for parsing S3, GS or local file urls and defining sound instrument types for metadata generation
 from typing import Tuple
 from urllib.parse import urlparse
+
+
+class InstrumentType:
+    NRS = "NRS"
+    ICLISTEN = "ICLISTEN"
+    SOUNDTRAP = "SOUNDTRAP"
 
 
 def parse_s3_or_gcp_url(url) -> Tuple[str, str, str]: